requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.10
astroid==2.4.2
billiard==3.3.0.23
celery==3.1.23
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
cryptography==3.2.1
defusedxml==0.7.0rc1
dj-database-url==0.4.1
Django==3.0.5
django-bootstrap3==7.0.1
django-cors-headers==3.5.0
django-crispy-forms==1.10.0
django-oauth-toolkit==1.3.3
django-phonenumber-field==5.0.0
django-rest-framework-social-oauth2==1.0.4
django-tables2==2.3.2
djangorestframework==3.12.1
djongo==1.3.3
dnspython==2.0.0
gunicorn==19.6.0
idna==2.10
isort==5.6.4
kombu==3.0.37
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
phonenumbers==8.12.13
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pymongo==3.11.1
PyMySQL==0.10.1
pyOpenSSL==19.1.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-core==3.3.3
sqlparse==0.2.4
toml==0.10.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
whitenoise==3.2.1
wrapt==1.12.1
